:recycle: Rename package from pipeline to models
