+++ models/processor_factory.go
@@ -1,4 +1,4 @@
-package pipeline
+package models
 
 import (
 	"fmt"
